refactor(streaming): split Pi subspawn tracker from drain coordinator
