+++ tests/unit/streaming/test_pi_quiescence.py
@@ -20,7 +20,7 @@ from meridian.lib.harness.connections.base import (
 )
 from meridian.lib.launch.launch_types import ResolvedLaunchSpec
 from meridian.lib.safety.permissions import UnsafeNoOpPermissionResolver
-from meridian.lib.streaming.pi_drain import PiSubspawnTracker
+from meridian.lib.streaming.pi_subspawn_tracker import PiSubspawnTracker
 from meridian.lib.streaming.spawn_manager import SpawnManager
 from tests.unit.streaming.pi_quiescence_test_helpers import (
     FakePiConnection as _FakePiConnection,
